Exchanged the database path in the constructor of the retry and batch operation templates for an instance of LiteDB (in order to allow thread-safe, concurrent operation of the LiteDB database).
Added new blocking operation, TakePendingRetry, that allows the caller to get blocked until a pending retry is available, allowing this way to prevent polling and implement an easier approach to process pending retries.

Slightly changed the meaning and renamed the complete operations in BatchOperationTemplate.

Documented the use of the library.